[UPDATE] Removed moduleinterface for model loading
