+++ engine/bindings/private/pathfinding/pathfinding_bindings.cpp
@@ -6,10 +6,9 @@
 
 namespace bindings
 {
-int32_t SetNavigationMesh(MAYBE_UNUSED PathfindingModule& self, MAYBE_UNUSED const std::string& path)
+int32_t SetNavigationMesh(PathfindingModule& self, const std::string& path)
 {
-    return 0;
-    // return self.SetNavigationMesh(path); // TODO: Not sure how to do this right now as we don't have access to model loading module, any ideas?
+    return self.SetNavigationMesh(path);
 }
 
 ComputedPath FindPath(PathfindingModule& self, const glm::vec3& start_pos, const glm::vec3& end_pos)
