+++ engine/model_loading/private/model_loading.cpp
@@ -1,4 +1,4 @@
-#include "model_loader.hpp"
+#include "model_loading.hpp"
 
 #include "animation.hpp"
 #include "batch_buffer.hpp"
@@ -13,6 +13,9 @@
 #include <glm/gtc/type_ptr.hpp>
 #include <glm/gtx/norm.hpp>
 #include <stb_image.h>
+
+constexpr static auto DEFAULT_LOAD_FLAGS = fastgltf::Options::DecomposeNodeMatrices | fastgltf::Options::LoadExternalBuffers | fastgltf::Options::LoadExternalImages;
+static fastgltf::Parser parser = fastgltf::Parser();
 
 namespace detail
 {
@@ -42,7 +45,7 @@ fastgltf::math::fmat4x4 ToFastGLTFMat4(const glm::mat4& glm_mat)
 }
 CPUModel ProcessModel(const fastgltf::Asset& gltf, const std::string_view name);
 
-CPUModel ModelLoader::LoadGLTF(std::string_view path)
+CPUModel ModelLoading::LoadGLTF(std::string_view path)
 {
     ZoneScoped;
 
@@ -66,7 +69,7 @@ CPUModel ModelLoader::LoadGLTF(std::string_view path)
         std::string zone = std::string(path) + " FastGLTF parse";
         ZoneName(zone.c_str(), 128);
 
-        auto loadedGltf = _parser.loadGltf(fileStream, directory, DEFAULT_LOAD_FLAGS);
+        auto loadedGltf = parser.loadGltf(fileStream, directory, DEFAULT_LOAD_FLAGS);
         if (!loadedGltf)
             throw std::runtime_error(getErrorMessage(loadedGltf.error()).data());
 
@@ -827,7 +830,7 @@ CPUModel ProcessModel(const fastgltf::Asset& gltf, const std::string_view name)
     return model;
 }
 
-void ModelLoader::ReadGeometrySizeGLTF(std::string_view path, uint32_t& vertexBufferSize, uint32_t& indexBufferSize)
+void ModelLoading::ReadGeometrySizeGLTF(std::string_view path, uint32_t& vertexBufferSize, uint32_t& indexBufferSize)
 {
     vertexBufferSize = 0;
     indexBufferSize = 0;
@@ -838,7 +841,7 @@ void ModelLoader::ReadGeometrySizeGLTF(std::string_view path, uint32_t& vertexBu
         throw std::runtime_error("Path not found!");
 
     std::string_view directory = path.substr(0, path.find_last_of('/'));
-    auto loadedGltf = _parser.loadGltf(fileStream, directory,
+    auto loadedGltf = parser.loadGltf(fileStream, directory,
         fastgltf::Options::DecomposeNodeMatrices | fastgltf::Options::LoadExternalBuffers | fastgltf::Options::LoadExternalImages);
 
     if (!loadedGltf)
