+++ engine/physics/private/systems/physics_system.cpp
@@ -11,11 +11,12 @@
 #include "glm/gtx/matrix_decompose.hpp"
 #include "graphics_context.hpp"
 #include "imgui.h"
-#include "model_loader.hpp"
+#include "model_loading.hpp"
 #include "renderer.hpp"
 #include "renderer_module.hpp"
 #include "resource_management/mesh_resource_manager.hpp"
 #include <Jolt/Physics/Collision/Shape/ScaledShape.h>
+#include <systems/physics_system.hpp>
 
 #include <tracy/Tracy.hpp>
 
@@ -24,11 +25,6 @@ PhysicsSystem::PhysicsSystem(Engine& engine, ECSModule& ecs, PhysicsModule& phys
     , _ecs(ecs)
     , _physicsModule(physicsModule)
 {
-    _collisionLoader = std::make_unique<ModelLoader>();
-}
-PhysicsSystem::~PhysicsSystem()
-{
-    _collisionLoader.reset();
 }
 
 entt::entity PhysicsSystem::LoadNodeRecursive(const CPUModel& models, ECSModule& ecs,
@@ -168,7 +164,7 @@ RigidbodyComponent PhysicsSystem::CreateMeshColliderBody(const CPUMesh<Vertex>& 
 
 void PhysicsSystem::CreateCollision(const std::string& path, const PhysicsShapes shapeType)
 {
-    CPUModel models = _collisionLoader.get()->LoadGLTF(path);
+    CPUModel models = ModelLoading::LoadGLTF(path);
     LoadNodeRecursive(models, _ecs, models.hierarchy.root, models.hierarchy, entt::null, shapeType);
 }
